+++ extension/logstash/lib/logstash/util/short_name_resolver.rb
